Use functions for generic standard models
